[FE,BE] Modify : 나의활동 좋아요 조회 기능 추가 / Like 엔티티 1개로 통합

+++ back/lab/src/main/java/com/mlp/lab/service/ShareRoomService.java
@@ -102,15 +102,15 @@ public class ShareRoomService {
         shareRoomRepository.save(shareRoom);
     }
 
-    public ShareRoomDto get(Integer roomNo) {
-        Optional<ShareRoom> result = shareRoomRepository.findById(roomNo);
+    public ShareRoomDto get(Long roomNo) {
+        Optional<ShareRoom> result = shareRoomRepository.findById(roomNo.intValue());
         ShareRoom shareRoom = result.orElseThrow();
         ShareRoomDto shareRoomDto = shareRoom.entityToDto(shareRoom);
         return shareRoomDto;
     }
 
-    public void remove(Integer roomNo) {
-        shareRoomRepository.deleteById(roomNo);
+    public void remove(Long roomNo) {
+        shareRoomRepository.deleteById(roomNo.intValue());
     }
 
     public void modify(ShareRoomDto shareRoomDto) { // 수정하기
@@ -186,14 +186,14 @@ public class ShareRoomService {
         return dtoList;
     }
 
-    public void increase(Integer roomNo) { // 좋아요 +1
+    public void increase(Long roomNo) { // 좋아요 +1
         Optional<ShareRoom> result = shareRoomRepository.findById(roomNo.intValue());
         ShareRoom shareRoom = result.orElseThrow();
         shareRoom.setRoomHit(shareRoom.getRoomHit()+1);
         shareRoomRepository.save(shareRoom);
     }
 
-    public void decrease(Integer roomNo) { // 좋아요 -1
+    public void decrease(Long roomNo) { // 좋아요 -1
         Optional<ShareRoom> result = shareRoomRepository.findById(roomNo.intValue());
         ShareRoom shareRoom = result.orElseThrow();
         shareRoom.setRoomHit(shareRoom.getRoomHit()-1);
